attached the ptroject member name lists
