+++ src/app/(legacy)/projects/page.tsx
@@ -1,94 +1,29 @@
-import projectsData from "@/data/projects.json";
+import { ProjectsGrid } from "@/components/ProjectsGrid";
 
 export const metadata = {
   title: "Projects",
-  description: "Explore the full Elixpo ecosystem — 11+ interconnected open-source projects spanning AI, web, and developer tools.",
+  description:
+    "Explore the Elixpo ecosystem — interconnected open-source projects spanning AI, web, and developer tools.",
   openGraph: {
     title: "Projects | Elixpo",
-    description: "Explore 11+ interconnected open-source projects spanning AI, web, and developer tools.",
+    description: "Interconnected open-source projects spanning AI, web, and developer tools.",
     images: ["/og-image.png"],
   },
 };
 
-function ProjectIcon({ icon, color }: { icon: string; color: string }) {
-  const icons: Record<string, React.ReactNode> = {
-    palette: <path d="M12 2C6.49 2 2 6.49 2 12s4.49 10 10 10c1.38 0 2.5-1.12 2.5-2.5 0-.61-.23-1.2-.64-1.67-.08-.1-.13-.21-.13-.33 0-.28.22-.5.5-.5H16c3.31 0 6-2.69 6-6 0-4.96-4.49-9-10-9zm-5.5 9c-.83 0-1.5-.67-1.5-1.5S5.67 8 6.5 8 8 8.67 8 9.5 7.33 11 6.5 11zm3-4C8.67 7 8 6.33 8 5.5S8.67 4 9.5 4s1.5.67 1.5 1.5S10.33 7 9.5 7zm5 0c-.83 0-1.5-.67-1.5-1.5S13.67 4 14.5 4s1.5.67 1.5 1.5S15.33 7 14.5 7zm3 4c-.83 0-1.5-.67-1.5-1.5S16.67 8 17.5 8s1.5.67 1.5 1.5-.67 1.5-1.5 1.5z" />,
-    search: <path d="M15.5 14h-.79l-.28-.27A6.471 6.471 0 0016 9.5 6.5 6.5 0 109.5 16c1.61 0 3.09-.59 4.23-1.57l.27.28v.79l5 4.99L20.49 19l-4.99-5zm-6 0C7.01 14 5 11.99 5 9.5S7.01 5 9.5 5 14 7.01 14 9.5 11.99 14 9.5 14z" />,
-    chat: <path d="M20 2H4c-1.1 0-2 .9-2 2v18l4-4h14c1.1 0 2-.9 2-2V4c0-1.1-.9-2-2-2zm0 14H6l-2 2V4h16v12z" />,
-    pencil: <path d="M3 17.25V21h3.75L17.81 9.94l-3.75-3.75L3 17.25zM20.71 7.04c.39-.39.39-1.02 0-1.41l-2.34-2.34c-.39-.39-1.02-.39-1.41 0l-1.83 1.83 3.75 3.75 1.83-1.83z" />,
-    document: <path d="M14 2H6c-1.1 0-2 .9-2 2v16c0 1.1.9 2 2 2h12c1.1 0 2-.9 2-2V8l-6-6zm4 18H6V4h7v5h5v11z" />,
-    bot: <path d="M20 9V7c0-1.1-.9-2-2-2h-3c0-1.66-1.34-3-3-3S9 3.34 9 5H6c-1.1 0-2 .9-2 2v2c-1.66 0-3 1.34-3 3s1.34 3 3 3v4c0 1.1.9 2 2 2h12c1.1 0 2-.9 2-2v-4c1.66 0 3-1.34 3-3s-1.34-3-3-3zM7.5 11.5c0-.83.67-1.5 1.5-1.5s1.5.67 1.5 1.5S9.83 13 9 13s-1.5-.67-1.5-1.5zM16 17H8v-2h8v2zm-1-4c-.83 0-1.5-.67-1.5-1.5S14.17 10 15 10s1.5.67 1.5 1.5S15.83 13 15 13z" />,
-    shield: <path d="M12 1L3 5v6c0 5.55 3.84 10.74 9 12 5.16-1.26 9-6.45 9-12V5l-9-4zm0 10.99h7c-.53 4.12-3.28 7.79-7 8.94V12H5V6.3l7-3.11v8.8z" />,
-    link: <path d="M3.9 12c0-1.71 1.39-3.1 3.1-3.1h4V7H7c-2.76 0-5 2.24-5 5s2.24 5 5 5h4v-1.9H7c-1.71 0-3.1-1.39-3.1-3.1zM8 13h8v-2H8v2zm9-6h-4v1.9h4c1.71 0 3.1 1.39 3.1 3.1s-1.39 3.1-3.1 3.1h-4V17h4c2.76 0 5-2.24 5-5s-2.24-5-5-5z" />,
-    user: <path d="M12 12c2.21 0 4-1.79 4-4s-1.79-4-4-4-4 1.79-4 4 1.79 4 4 4zm0 2c-2.67 0-8 1.34-8 4v2h16v-2c0-2.66-5.33-4-8-4z" />,
-    clock: <path d="M11.99 2C6.47 2 2 6.48 2 12s4.47 10 9.99 10C17.52 22 22 17.52 22 12S17.52 2 11.99 2zM12 20c-4.42 0-8-3.58-8-8s3.58-8 8-8 8 3.58 8 8-3.58 8-8 8zm.5-13H11v6l5.25 3.15.75-1.23-4.5-2.67z" />,
-    file: <path d="M14 2H6c-1.1 0-2 .9-2 2v16c0 1.1.9 2 2 2h12c1.1 0 2-.9 2-2V8l-6-6zm-1 7V3.5L18.5 9H13z" />,
-  };
-
-  return (
-    <div className="w-12 h-12 rounded-xl flex items-center justify-center" style={{ backgroundColor: color + "15" }}>
-      <svg className="w-6 h-6" style={{ color }} fill="currentColor" viewBox="0 0 24 24">
-        {icons[icon] || null}
-      </svg>
-    </div>
-  );
-}
-
 export default function ProjectsPage() {
   return (
-    <section className="py-24 md:py-32">
-      <div className="max-w-7xl mx-auto px-6">
-        <div className="max-w-3xl mb-16">
-          <h1 className="text-5xl md:text-7xl font-bold tracking-tight mb-6">Projects</h1>
-          <p className="text-lg text-muted">
-            The full Elixpo ecosystem — {projectsData.projects.length} interconnected open-source projects
-            spanning AI, web, and developer tools.
-          </p>
-        </div>
-
-        <div className="space-y-6">
-          {projectsData.projects.map((project) => (
-            <a
-              key={project.name}
-              href={project.url}
-              target={project.url !== "#" ? "_blank" : undefined}
-              rel={project.url !== "#" ? "noopener noreferrer" : undefined}
-              className={`group block rounded-3xl bg-card border border-border p-8 transition-all duration-300 ${
-                project.url !== "#" ? "hover:shadow-lg hover:border-transparent" : ""
-              }`}
-            >
-              <div className="flex flex-col md:flex-row md:items-start gap-6">
-                <ProjectIcon icon={project.icon} color={project.color} />
-                <div className="flex-1">
-                  <div className="flex items-center gap-3 mb-2">
-                    <h2 className="text-xl font-semibold group-hover:text-accent transition-colors">
-                      {project.name}
-                    </h2>
-                    {project.url !== "#" && (
-                      <span className="text-xs px-2 py-0.5 rounded-full bg-green-100 text-green-700 font-medium">
-                        Live
-                      </span>
-                    )}
-                  </div>
-                  <p className="text-sm text-accent font-medium mb-2">{project.tagline}</p>
-                  <p className="text-sm text-muted leading-relaxed mb-4">{project.description}</p>
-                  <div className="flex flex-wrap gap-1.5">
-                    {project.tech.map((t) => (
-                      <span key={t} className="text-[11px] px-2.5 py-1 rounded-full bg-card-hover text-muted font-medium">
-                        {t}
-                      </span>
-                    ))}
-                  </div>
-                </div>
-                {project.url !== "#" && (
-                  <svg className="w-5 h-5 text-muted group-hover:text-accent transition-colors shrink-0 hidden md:block" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M10 6H6a2 2 0 00-2 2v10a2 2 0 002 2h10a2 2 0 002-2v-4M14 4h6m0 0v6m0-6L10 14" />
-                  </svg>
-                )}
-              </div>
-            </a>
-          ))}
-        </div>
+    <section className="bg-black text-[#E1E0CC] py-16 px-4 sm:px-6 lg:px-8 min-h-screen select-none">
+      <div className="max-w-7xl mx-auto">
+        <header className="mb-12 sm:mb-16">
+          <span className="text-[10px] uppercase tracking-widest font-mono text-primary/80 block mb-3">
+            The Elixpo Ecosystem
+          </span>
+          <h1 className="text-4xl sm:text-5xl md:text-6xl font-serif italic text-white tracking-tight">
+            Projects
+          </h1>
+        </header>
+        <ProjectsGrid />
       </div>
     </section>
   );
